Changed a Lot

+++ src/main/java/invasion/nexus/SpawnPoint.java
@@ -6,10 +6,10 @@ import net.minecraft.util.math.BlockPos;
 
 public class SpawnPoint implements IPolarAngle, Comparable<IPolarAngle> {
     private final BlockPos pos;
-    private final int spawnAngle;
+    private final float spawnAngle;
     private final SpawnType spawnType;
 
-    public SpawnPoint(int x, int y, int z, int angle, SpawnType type) {
+    public SpawnPoint(int x, int y, int z, float angle, SpawnType type) {
         pos = new BlockPos(x, y, z);
         spawnAngle = angle;
         spawnType = type;
@@ -20,7 +20,7 @@ public class SpawnPoint implements IPolarAngle, Comparable<IPolarAngle> {
     }
 
     @Override
-    public int getAngle() {
+    public float getAngle() {
         return spawnAngle;
     }
 
@@ -30,8 +30,7 @@ public class SpawnPoint implements IPolarAngle, Comparable<IPolarAngle> {
 
     @Override
     public int compareTo(IPolarAngle polarAngle) {
-        return Integer.compare(spawnAngle, polarAngle.getAngle());
-
+        return Float.compare(spawnAngle, polarAngle.getAngle());
     }
 
     @Override
